+++ src/poset/badger_store_test.go
@@ -150,17 +150,17 @@ func TestLoadBadgerStore(t *testing.T) {
 
 func TestDBEventMethods(t *testing.T) {
 	cacheSize := 0
-	testSize := 100
+	testSize := int64(100)
 	store, participants := initBadgerStore(cacheSize, t)
 	defer removeBadgerStore(store, t)
 
 	//insert events in db directly
 	events := make(map[string][]Event)
-	topologicalIndex := 0
+	topologicalIndex := int64(0)
 	var topologicalEvents []Event
 	for _, p := range participants {
 		var items []Event
-		for k := 0; k < testSize; k++ {
+		for k := int64(0); k < testSize; k++ {
 			event := NewEvent(
 				[][]byte{[]byte(fmt.Sprintf("%s_%d", p.hex[:5], k))},
 				[]InternalTransaction{},
@@ -235,13 +235,13 @@ func TestDBEventMethods(t *testing.T) {
 	}
 
 	//check that participant events where correctly added
-	skipIndex := -1 //do not skip any indexes
+	skipIndex := int64(-1) //do not skip any indexes
 	for _, p := range participants {
 		pEvents, err := store.dbParticipantEvents(p.hex, skipIndex)
 		if err != nil {
 			t.Fatal(err)
 		}
-		if l := len(pEvents); l != testSize {
+		if l := int64(len(pEvents)); l != testSize {
 			t.Fatalf("%s should have %d events, not %d", p.hex, testSize, l)
 		}
 
@@ -328,8 +328,8 @@ func TestDBBlockMethods(t *testing.T) {
 	store, participants := initBadgerStore(cacheSize, t)
 	defer removeBadgerStore(store, t)
 
-	index := 0
-	roundReceived := 5
+	index := int64(0)
+	roundReceived := int64(5)
 	transactions := [][]byte{
 		[]byte("tx1"),
 		[]byte("tx2"),
@@ -411,7 +411,7 @@ func TestDBFrameMethods(t *testing.T) {
 		event.Sign(p.privKey)
 		events = append(events, event)
 
-		root := NewBaseRoot(id)
+		root := NewBaseRoot(int64(id))
 		roots = append(roots, root)
 	}
 	frame := Frame{
@@ -443,7 +443,7 @@ func TestDBFrameMethods(t *testing.T) {
 func TestBadgerEvents(t *testing.T) {
 	//Insert more events than can fit in cache to test retrieving from db.
 	cacheSize := 10
-	testSize := 100
+	testSize := int64(100)
 	store, participants := initBadgerStore(cacheSize, t)
 	defer removeBadgerStore(store, t)
 
@@ -451,7 +451,7 @@ func TestBadgerEvents(t *testing.T) {
 	events := make(map[string][]Event)
 	for _, p := range participants {
 		var items []Event
-		for k := 0; k < testSize; k++ {
+		for k := int64(0); k < testSize; k++ {
 			event := NewEvent([][]byte{[]byte(fmt.Sprintf("%s_%d", p.hex[:5], k))},
 				[]InternalTransaction{},
 				[]BlockSignature{{Validator: []byte("validator"), Index: 0, Signature: "r|s"}},
@@ -484,13 +484,13 @@ func TestBadgerEvents(t *testing.T) {
 	}
 
 	//check retrieving events per participant
-	skipIndex := -1 //do not skip any indexes
+	skipIndex := int64(-1) //do not skip any indexes
 	for _, p := range participants {
 		pEvents, err := store.ParticipantEvents(p.hex, skipIndex)
 		if err != nil {
 			t.Fatal(err)
 		}
-		if l := len(pEvents); l != testSize {
+		if l := int64(len(pEvents)); l != testSize {
 			t.Fatalf("%s should have %d events, not %d", p.hex, testSize, l)
 		}
 
@@ -517,7 +517,7 @@ func TestBadgerEvents(t *testing.T) {
 		}
 	}
 
-	expectedKnown := make(map[int]int)
+	expectedKnown := make(map[int64]int64)
 	for _, p := range participants {
 		expectedKnown[p.id] = testSize - 1
 	}
@@ -589,8 +589,8 @@ func TestBadgerBlocks(t *testing.T) {
 	store, participants := initBadgerStore(cacheSize, t)
 	defer removeBadgerStore(store, t)
 
-	index := 0
-	roundReceived := 5
+	index := int64(0)
+	roundReceived := int64(5)
 	transactions := [][]byte{
 		[]byte("tx1"),
 		[]byte("tx2"),
@@ -671,7 +671,7 @@ func TestBadgerFrames(t *testing.T) {
 		event.Sign(p.privKey)
 		events = append(events, event)
 
-		root := NewBaseRoot(id)
+		root := NewBaseRoot(int64(id))
 		roots = append(roots, root)
 	}
 	frame := Frame{
